Finished 5.3

+++ src/main/java/com/fi0x/cc/exercises/week3/Exercise3.java
@@ -38,7 +38,7 @@ public class Exercise3 extends PApplet
         float pmY = pmouseY - centerY;
 
         translate(centerX, centerY);
-        for (float i = 0; i < 6; i++) {
+        for (int i = 0; i < 6; i++) {
             rotate(radians(360 / 6f));
             line(mX, mY, pmX, pmY);
             pushMatrix();
